Sun, Mar 10, 2024, 6:36 PM +00:00

+++ CS50Python/pset4/bitcoin/bitcoin.py
@@ -3,7 +3,7 @@ import requests
 import sys
 
 try:
-    n = (float)sys.argv[1]
+    n = (int)sys.argv[1]
 except ValueError:
     sys.exit("Not a valid value")
